+++ src/cenreg/model/__init__.py
@@ -1,3 +1,1 @@
 from .cjd2F_np import cjd2surv
-
-__all__ = ["cjd2surv"]